// web/src/api/generated/models/DashboardModulePreferenceList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DashboardModulePreferenceView } from './DashboardModulePreferenceView';
import {
    DashboardModulePreferenceViewFromJSON,
    DashboardModulePreferenceViewFromJSONTyped,
    DashboardModulePreferenceViewToJSON,
    DashboardModulePreferenceViewToJSONTyped,
} from './DashboardModulePreferenceView';

/**
 * 
 * @export
 * @interface DashboardModulePreferenceList
 */
export interface DashboardModulePreferenceList {
    /**
     * 
     * @type {Array<DashboardModulePreferenceView>}
     * @memberof DashboardModulePreferenceList
     */
    items: Array<DashboardModulePreferenceView>;
}

/**
 * Check if a given object implements the DashboardModulePreferenceList interface.
 */
export function instanceOfDashboardModulePreferenceList(value: object): value is DashboardModulePreferenceList {
    if (!('items' in value) || value['items'] === undefined) return false;
    return true;
}

export function DashboardModulePreferenceListFromJSON(json: any): DashboardModulePreferenceList {
    return DashboardModulePreferenceListFromJSONTyped(json, false);
}

export function DashboardModulePreferenceListFromJSONTyped(json: any, ignoreDiscriminator: boolean): DashboardModulePreferenceList {
    if (json == null) {
        return json;
    }
    return {
        
        'items': ((json['items'] as Array<any>).map(DashboardModulePreferenceViewFromJSON)),
    };
}

export function DashboardModulePreferenceListToJSON(json: any): DashboardModulePreferenceList {
    return DashboardModulePreferenceListToJSONTyped(json, false);
}

export function DashboardModulePreferenceListToJSONTyped(value?: DashboardModulePreferenceList | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'items': ((value['items'] as Array<any>).map(DashboardModulePreferenceViewToJSON)),
    };
}